split spiders
